BOCM-52: pr amends - refactored to move makeDownloadLinksDataForTemplate function into helpers file
